requirements.txt
numpy
pandas
pydantic
python-dotenv
openai
anthropic
google-genai
lxml
scipy
statsmodels
